User Changes - Task Changes
Some minor code cleanup
Add check to see if user is active
Added ability to verify, admin or activate a user(or disable these)
Added ability to open and close task.
Added finished value to task model.
Updated templates for changes.

Added requirement for flask-migrate as I plan to use this in the future. Currently not implemented.